+++ include/libft/srcs/ft_swap_f.c
@@ -6,15 +6,15 @@
 /*   By: fmanetti <fmanetti@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/05/24 17:27:20 by fmanetti          #+#    #+#             */
-/*   Updated: 2020/05/24 17:31:30 by fmanetti         ###   ########.fr       */
+/*   Updated: 2020/09/02 13:26:14 by fmanetti         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../libft.h"
 
-void	ft_swap_f(float *a, float *b)
+void	ft_swap_f(double *a, double *b)
 {
-	float tmp;
+	double tmp;
 
 	tmp = *a;
 	*a = *b;
