+++ include/struct.h
@@ -6,7 +6,7 @@
 /*   By: fmanetti <fmanetti@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/05/18 14:05:19 by fmanetti          #+#    #+#             */
-/*   Updated: 2020/08/19 19:56:54 by fmanetti         ###   ########.fr       */
+/*   Updated: 2020/09/02 13:23:38 by fmanetti         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@ typedef unsigned int t_uint;
 
 typedef struct		s_point
 {
-	float	x;
-	float	y;
-	float	z;
+	double	x;
+	double	y;
+	double	z;
 }					t_point;
 
 typedef struct		s_ray
@@ -28,8 +28,8 @@ typedef struct		s_ray
 
 typedef	struct		s_coord
 {
-	float	x1;
-	float	x2;
+	double	x1;
+	double	x2;
 }					t_coord;
 
 typedef	struct 		s_color
@@ -54,25 +54,25 @@ typedef struct 		s_setting
 {
 	unsigned int	width;
 	unsigned int	heigth;
-	float			amblrat; //luminance ratio 
+	double			amblrat; //luminance ratio 
 	t_color			amblclr;
-	float			scale;
-	float			imageAR;
+	double			scale;
+	double			imageAR;
 	void			*head;
 	t_mlx			d;
 }					t_setting;
 
 typedef struct		s_coeff
 {
-	float a;
-	float b;
-	float c;
+	double a;
+	double b;
+	double c;
 }					t_coeff;
 
 typedef struct		s_cobar
 {
-	float u;
-	float v;
+	double u;
+	double v;
 	t_point pvec;
 	t_point qvec;
 	t_point tvec;
